compilers: split pgi compiler out of compilers module
